feat(SocialLinks): Remove unused social links
Removes the Briefcase and ListMusic icons from the SocialLinks
component as they are not being used in the current implementation.
This helps to keep the codebase clean and focused on the
essential features.

fix(HomePage): Disable currently working status

Disables the `isCurrentlyWorking` flag in the HomePage component as
this functionality is not currently being used. This change helps to
simplify the component and remove unused code.