feat(social-links): Update LinkedIn URL
Changes the LinkedIn URL in the `SocialLinks` and `OtherLinks` components to a new URL.

This change updates the LinkedIn URL to a new profile, which is the correct URL for the portfolio owner.

feat(button): Improve tooltip styles

Updates the styles of the tooltip component in the `Button` component. The changes include:

- Updating the background and text colors to use the `neutral-0` and `neutral-100` colors, which are part of the design system.
- Adjusting the positioning and appearance of the tooltip arrow.

feat(home-page): Replace Footer with Calls to Action

Replaces the `Footer` component with a new `CallsToAction` component on the home page. This change better aligns with the overall design and functionality of the portfolio website.

feat(project-card): Add tooltip to buttons

Adds a `tooltip` prop to the `Button` component used in the `ProjectCard` component. This provides an additional accessibility feature by displaying a tooltip when the button is hovered over.

feat(tooltip): Improve tooltip visibility

Enhances the `Tooltip` component by adding additional logging and state management to improve the visibility and behavior of the tooltip. This includes adding console logs to help with debugging and updating the `useEffect` hook to log changes in the `isVisible` state.

feat(other-links): Add tooltip to buttons

Adds a `tooltip` prop to the `FilledButton` component used in the `OtherLinks` component. This provides an additional accessibility feature by displaying a tooltip when the button is hovered over.